# perf: avoid Headers initialization