## Rationale

<!-- Briefly explain the purpose of this pull request, if not already
justifiable with the above section. If it is, you may omit this section. -->

Improves performance by avoiding the initialization process of `Headers`.

## Benchmark

```js
// bench.js
const { Request, Response } = require(".");

const benchmark = require("benchmark");

const suite = new benchmark.Suite();

const input = "https://example.com/post";

suite.add("new Request()", () => {
  new Request(input, undefined);
});

suite.add("new Response()", () => {
  new Response(null, undefined);
});

suite.on("cycle", function (event) {
  console.log(String(event.target));
});

// run async
new Promise((resolve) => setTimeout(resolve, 7000)).then(() =>
  suite.run({ async: true })
);

```

- *main*
```
new Request() x 100,116 ops/sec ±4.53% (74 runs sampled)
new Response() x 1,922,146 ops/sec ±2.12% (83 runs sampled)
```

- *this patch*
```
new Request() x 106,769 ops/sec ±3.77% (80 runs sampled)
new Response() x 2,330,854 ops/sec ±1.48% (89 runs sampled)
```