@tsctx 2023-11-27
Hi @KhafraDev @ronag @mcollina,
What do you think about this?